chore: all-in-one index.js
