+++ package.json
@@ -20,7 +20,7 @@
     "api"
   ],
   "bin": {
-    "alola": "./index.js"
+    "alola": "./alola.js"
   },
   "devDependencies": {
     "prettier": "2.7.1"
